+++ kernel/emulation/linux/misc/ioctl.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_IOCTL_H
 #define LINUX_IOCTL_H
 
-int sys_ioctl(int fd, int req, void* arg);
+long sys_ioctl(int fd, int req, void* arg);
 
 #endif
