+++ kernel/emulation/linux/network/connect.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_CONNECT_H
 #define LINUX_CONNECT_H
 
-int sys_connect(int fd, const void* name, int socklen);
+long sys_connect(int fd, const void* name, int socklen);
 
 #define LINUX_SYS_CONNECT	2
