+++ kernel/emulation/linux/signal/kill.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_KILL_H
 #define LINUX_KILL_H
 
-int sys_kill(int pid, int signum, int posix);
+long sys_kill(int pid, int signum, int posix);
 
 #endif
