+++ kernel/emulation/linux/unistd/access.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_ACCESS_H
 #define LINUX_ACCESS_H
 
-int sys_access(const char* filename, int amode);
+long sys_access(const char* filename, int amode);
 
 #define LINUX_AT_FDCWD -100
