+++ kernel/emulation/linux/unistd/dup.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_DUP_H
 #define LINUX_DUP_H
 
-int sys_dup(int fd);
+long sys_dup(int fd);
 
 #endif
