+++ kernel/emulation/linux/unistd/dup2.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_DUP2_H
 #define LINUX_DUP2_H
 
-int sys_dup2(int fd_from, int fd_to);
+long sys_dup2(int fd_from, int fd_to);
 
 #endif
