+++ kernel/emulation/linux/unistd/exit.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_EXIT_H
 #define LINUX_EXIT_H
 
-int sys_exit(int status);
+long sys_exit(int status);
 
 #endif
