+++ kernel/emulation/linux/unistd/fchdir.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_FCHDIR_H
 #define LINUX_FCHDIR_H
 
-int sys_fchdir(int fd);
+long sys_fchdir(int fd);
 
 #endif
