+++ kernel/emulation/linux/unistd/fchmod.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_FCHMOD_H
 #define LINUX_FCHMOD_H
 
-int sys_fchmod(int fd, int mode);
+long sys_fchmod(int fd, int mode);
 
 #endif
