+++ kernel/emulation/linux/unistd/geteuid.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_GETEUID_H
 #define LINUX_GETEUID_H
 
-int sys_geteuid(void);
+long sys_geteuid(void);
 
 #endif
