+++ kernel/emulation/linux/unistd/getuid.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_GETUID_H
 #define LINUX_GETUID_H
 
-int sys_getuid(void);
+long sys_getuid(void);
 
 #endif
