+++ kernel/emulation/linux/unistd/setegid.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_SETEGID_H
 #define LINUX_SETEGID_H
 
-int sys_setegid(int egid);
+long sys_setegid(int egid);
 
 #endif
