+++ kernel/emulation/linux/unistd/seteuid.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_SETEUID_H
 #define LINUX_SETEUID_H
 
-int sys_seteuid(int euid);
+long sys_seteuid(int euid);
 
 #endif
