+++ kernel/emulation/linux/unistd/setgid.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_SETGID_H
 #define LINUX_SETGID_H
 
-int sys_setgid(int gid);
+long sys_setgid(int gid);
 
 #endif
