+++ kernel/emulation/linux/unistd/setsid.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_SETSID_H
 #define LINUX_SETSID_H
 
-int sys_setsid(void);
+long sys_setsid(void);
 
 #endif
