+++ kernel/emulation/linux/unistd/sync.h
@@ -1,7 +1,7 @@
 #ifndef LINUX_SYNC_H
 #define LINUX_SYNC_H
 
-int sys_sync(void);
+long sys_sync(void);
 
 #endif
